% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_danger.R
\name{fire_danger}
\alias{fire_danger}
\title{fire_danger}
\usage{
fire_danger(fire_df)
}
\arguments{
\item{fire_df}{A dataframe containing columns temp (C), humidity (RH), wind_speed (mph), and fuel_moisture (Live fuel moisture content)}
}
\value{
fire_score   =  Fire danger index score

fire_danger_level   =  Fire danger level
}
\description{
Gives the fire danger index score and fire danger level based on temperature, humidity, wind speed, live fuel moisture content.
}
\author{
Allison Hacker
}
