% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_records.R
\name{heat_records}
\alias{heat_records}
\title{heat_records}
\usage{
heat_records(temps, records)
}
\arguments{
\item{records}{Record high temperature}

\item{temp}{Average daily temperature}
}
\value{
record_high = number of days that exceeded the previous record high temperature
}
\description{
Gives the number of days where the temperature exceeded the previous temperature record.
}
\author{
Allison Hacker
}
