% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_summary.R
\name{precip_summary}
\alias{precip_summary}
\title{precip_summary}
\usage{
precip_summary(amt, type, snow_density = 0.312)
}
\arguments{
\item{type}{Precipitation type (rain or snow)}

\item{snow_density}{Desnity of snow (kg/m2)}

\item{amount}{Precipitation amount (cm)}
}
\value{
min_precip = minimum daily precipitation (cm)

max_precip = maximum daily precipitation (cm)

mean_precip = mean daily precipitation (cm)

snowdays = number of days with snow

total_swe = snow water equivalent (cm/day)
}
\description{
Summarizes the minimum, maximum, and mean amount of daily precipitation. Also gives the number snow days and the total snow water equivalent (SWE)
}
\author{
Allison Hacker
}
